V1.1

+++ ModelLibrary/DTOs/Algorithm/AlgorithmRawOutput.cs
@@ -2,12 +2,30 @@
 {
     public class AlgorithmRawOutput
     {
-        public int TotalSalary { get; set; } // Tong chi phi toi uu
-        public int TotalExper { get; set; } // Tong chat luong du an
-        public int TimeFinish { get; set; }
+        public long TotalSalary
+        {
+            get; set;
+        } // Tong chi phi toi uu
+        public int TotalExper
+        {
+            get; set;
+        } // Tong chat luong du an
+        public int TimeFinish
+        {
+            get; set;
+        }
 
-        public int[]? TaskBegin { get; set; }
-        public int[]? TaskFinish { get; set; }
-        public int[]? Genes { get; set; }
+        public int[]? TaskBegin
+        {
+            get; set;
+        }
+        public int[]? TaskFinish
+        {
+            get; set;
+        }
+        public int[]? Genes
+        {
+            get; set;
+        }
     }
 }
