+++ ModelLibrary/DTOs/Algorithm/EquipmentInputToORDTO.cs
@@ -2,8 +2,17 @@
 {
     public class EquipmentInputToORDTO
     {
-        public int Id { get; set; }
-        public List<FunctionInputToORDTO> Functions { get; set; }
-        public int UnitPrice { get; set; }
+        public int Id
+        {
+            get; set;
+        }
+        public List<FunctionInputToORDTO> Functions
+        {
+            get; set;
+        }
+        public long UnitPrice
+        {
+            get; set;
+        }
     }
 }
