+++ ModelLibrary/DTOs/Algorithm/OutputFromORDTO.cs
@@ -4,11 +4,23 @@ namespace ModelLibrary.DTOs.Algorithm
 {
     public class OutputFromORDTO
     {
-        public int totalSalary { get; set; } // Tong chi phi toi uu
-        public int totalExper { get; set; } // Tong chat luong du an
-        public int timeFinish { get; set; }
+        public long totalSalary
+        {
+            get; set;
+        } // Tong chi phi toi uu
+        public int totalExper
+        {
+            get; set;
+        } // Tong chat luong du an
+        public int timeFinish
+        {
+            get; set;
+        }
 
-        public List<TaskScheduleResultDTO>? tasks { get; set; }
+        public List<TaskScheduleResultDTO>? tasks
+        {
+            get; set;
+        }
 
         public OutputFromORDTO()
         {
